Finish test suite and fix issues with commentsController

+++ resources/comments/commentsController.js
@@ -16,7 +16,9 @@ exports.createPhotoComment = async (req, res) => {
       .status(400)
       .json({ message: 'userId was not attached to the req.body' });
   } else if (!req.body.text) {
-    return res.status(400).json({ message: 'text was not attached to the req.body' });
+    return res
+      .status(400)
+      .json({ message: 'text was not attached to the req.body' });
   } else if (!req.body.username) {
     return res
       .status(400)
@@ -30,27 +32,34 @@ exports.createPhotoComment = async (req, res) => {
         userId: req.body.userId,
         text: req.body.text,
         username: req.body.username
-      }
-
-      console.log(comment);
+      };
 
       const [id] = await go.createOne('comments', 'id', comment);
       const data = await go.getById('comments', id);
-      return res.status(201).json({ message: 'Comment successfully created!', data });
+      return res
+        .status(201)
+        .json({ message: 'Comment successfully created!', data });
+    } else {
+      return res
+        .status(401)
+        .json({
+          message:
+            'Unauthorized: You are not authorized to create a comment for someone else'
+        });
     }
-    else {
-      return res.status(401).json({ message: 'Unauthorized: You are not authorized to create a comment for someone else' });
-    }
-
   } catch (error) {
     console.error(error);
-    return res.status(400).json({ message: "Couldn't create comment", error: error });
+    return res
+      .status(400)
+      .json({ message: "Couldn't create comment", error: error });
   }
 };
 
 exports.getCommentsByImageId = async (req, res) => {
   if (!req.params.id) {
-    return res.status(400).json({ message: 'id was not attached to the req.params' });
+    return res
+      .status(400)
+      .json({ message: 'id was not attached to the req.params' });
   }
   const { id } = req.params;
   try {
@@ -74,7 +83,7 @@ exports.getCommentsByImageId = async (req, res) => {
     return res.status(200).json(data);
   } catch (err) {
     console.error(error);
-    res
+    return res
       .status(400)
       .json({ message: "Couldn't find the photo's comments", error: error });
   }
@@ -85,17 +94,19 @@ exports.getCommentsByImageId = async (req, res) => {
 
 exports.createProjectComment = async (req, res) => {
   if (!req.body.projectId) {
-    res
+    return res
       .status(400)
       .json({ message: 'projectId was not attached to the req.body' });
   } else if (!req.body.userId) {
-    res
+    return res
       .status(400)
       .json({ message: 'userId was not attached to the req.body' });
   } else if (!req.body.text) {
-    return res.status(400).json({ message: 'text was not attached to the req.body' });
+    return res
+      .status(400)
+      .json({ message: 'text was not attached to the req.body' });
   } else if (!req.body.username) {
-    res
+    return res
       .status(400)
       .json({ message: 'username was not attached to the req.body' });
   }
@@ -104,21 +115,30 @@ exports.createProjectComment = async (req, res) => {
     if (await userMatches(req.user, req.body.userId)) {
       const [id] = await go.createOne('comments', 'id', req.body);
       const data = await go.getById('comments', id);
-      return res.status(201).json({ message: 'Comment successfully created!', data });
+      return res
+        .status(201)
+        .json({ message: 'Comment successfully created!', data });
+    } else {
+      return res
+        .status(401)
+        .json({
+          message:
+            'Unauthorized: You are not authorized to create a comment for someone else'
+        });
     }
-    else {
-      return res.status(401).json({ message: 'Unauthorized: You are not authorized to create a comment for someone else' });
-    }
-
   } catch (error) {
     console.error(error);
-    return res.status(400).json({ message: "Couldn't create comment", error: error });
+    return res
+      .status(400)
+      .json({ message: "Couldn't create comment", error: error });
   }
 };
 
 exports.getCommentsByProjectId = async (req, res) => {
   if (!req.params.id) {
-    return res.status(400).json({ message: 'id was not attached to the req.params' });
+    return res
+      .status(400)
+      .json({ message: 'id was not attached to the req.params' });
   }
 
   const { id } = req.params;
@@ -142,7 +162,7 @@ exports.getCommentsByProjectId = async (req, res) => {
     return res.status(200).json(data);
   } catch (err) {
     console.error(error);
-    res
+    return res
       .status(400)
       .json({ message: "Couldn't find the project's comments", error: error });
   }
@@ -160,22 +180,43 @@ exports.updateCommentById = async (req, res) => {
       await go.updateById('comments', req.body, id);
       const data = await go.getById('comments', id);
       return res.status(200).json(data);
+    } else {
+      return res
+        .status(401)
+        .json({
+          message:
+            "Unauthorized: You may not update comments that don't belong to you."
+        });
     }
-    else {
-      return res.status(401).json({ message: "Unauthorized: You may not update comments that don't belong to you." });
-    }
-
   } catch (error) {
-    return res.status(400).json({ message: "Couldn't update comment.", error: error });
+    return res
+      .status(400)
+      .json({ message: "Couldn't update comment.", error: error });
   }
 };
 
-exports.deleteProjectById = async (req, res) => {
+exports.deleteCommentById = async (req, res) => {
   const { id } = req.params;
   try {
-    await go.destroyById('comments', id);
-    return res.status(200).json({ message: 'Comment successfully deleted' });
+    const comment = await go.getById('comments', id);
+
+    if (comment.length === 0) {
+      return res
+        .status(404)
+        .json({ message: 'A comment with that ID does not exist.' });
+    }
+
+    if (await userMatches(req.user, comment.userId)) {
+      await go.destroyById('comments', id);
+      return res.status(200).json({ message: 'Comment successfully deleted' });
+    } else {
+      return res
+        .status(401)
+        .json({ message: 'Unauthorized: You may not delete this comment.' });
+    }
   } catch (error) {
-    return res.status(400).json({ message: "Couldn't delete comment.", error: error });
+    return res
+      .status(400)
+      .json({ message: "Couldn't delete comment.", error: error });
   }
 };
